requirements.txt
pandas==1.2.2
Pillow==8.1.0
streamlit==0.62.1
xlrd==2.0.1
openpyxl==3.0.6